requirements.txt
fyers-apiv3
python-dotenv
plotly
pandas